requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
Pillow==9.4.0
PyAudio==0.2.13
pydicom==2.3.1
Werkzeug==2.2.2
